* win2k/IMDisplay/res/{IMDisplay.ico, IMDisplayDoc.ico}: Replaced with GraphicsMagick icon prepared by Jaroslav Fojtik.
* coders/svg.c (ReadSVGImage): Use runtime initialization of
SAXModules rather than static initialization.

* magick/command.c: Commands now support reading an image from
stdin in conjunction with a subrange specification (e.g. "-[1]").
Problem was reported by Mario Becroft.

* magick/common.h: New header file to incorporate the common bits
shared by studio.h and api.h.

* ltdl/ltdl.c: Update libltdl to 2.2.6b in order to fix security
issue.  Resolves CVE-2009-3736 as it pertains to GraphicsMagick.

+++ coders/topol.c
@@ -570,7 +570,7 @@ NoMEZ:		/*Clean up palette and clone_info*/
 	    j = i; /* MEZ[i];	ignore MEZ!!! proper palete element after reindexing */	
           else
             j = MEZ[i];			/* ?? I do not know, big mess ?? */
-	  if(j>=image->colors) j=image->colors-1;
+	  if(j>=(long) image->colors) j=image->colors-1;
           image->colormap[j].red = ScaleCharToQuantum(ReadBlobByte(palette));
           image->colormap[j].green = ScaleCharToQuantum(ReadBlobByte(palette));
           image->colormap[j].blue = ScaleCharToQuantum(ReadBlobByte(palette));
